Here's a cmake file for cross-platform builds
This uses the "cmake" utility, which claims to make makefiles for
a variety of platforms.  (cmake is an alternative to autotools, see
cmake.org for details.)

You use it like this:  

    cmake .  # you only have to do this once

    make -f Makefile
    (hack, hack)
    make -f Makefile
    (repeat)

The "-f Makefile" is needed because we have a file named "makefile",
and I want to make sure you get the new one that cmake made for you.
If we decide to switch to cmake we can remove our old makefile and the
-f requirement will go away.

I've verified that this works on Ubuntu Hardy, Karmic, and Lucid (yay,
Lucid!).  I'm curious if it works on Windows and Mac OSX.