first MillingCutter::offsetCutter(double d) implementation
c2 = c1.offsetCutter(d)
will return a new cutter c2 which is offset by d from the original cutter c1

this can be useful for a "stock-to-leave" setting in upstream higher-level operations. it is up to the higher level algorithm to do any correction in the z-coordinate which the offsetCutter path may require.